backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
openai==1.47.0
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
sentence-transformers==3.1.1
chromadb==0.5.5
requests==2.32.3
PyPDF2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
Pillow==10.4.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.2
deep-translator==1.11.4
transformers==4.45.1
torch==2.4.1
google-auth
google-auth-oauthlib
google-auth-httplib2
requests==2.32.3
opencv-contrib-python==4.9.0.80
google-api-python-client